+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/ViewModels/Maintenance/ExchangeCreateVM.cs
@@ -69,6 +69,20 @@ namespace TD.WPF.ViewModels.Maintenance
             }
         }
 
+        public string CountryCode
+        {
+            get { return _exchange.CountryCode; }
+            set
+            {
+                if (_exchange.CountryCode != value)
+                {
+                    _exchange.CountryCode = value;
+                    RaisePropertyChanged(nameof(CountryCode));
+                    ValidateProperty(_exchange, nameof(CountryCode));
+                }
+            }
+        }
+
         public int DefaultCurrencyId
         {
             get { return _exchange.DefaultCurrencyId; }
@@ -93,6 +107,34 @@ namespace TD.WPF.ViewModels.Maintenance
                     _exchange.DefaultCurrency = value;
                     RaisePropertyChanged(nameof(DefaultCurrency));
                     ValidateProperty(_exchange, nameof(DefaultCurrency));
+                }
+            }
+        }
+
+        public string Timezone
+        {
+            get { return _exchange.Timezone; }
+            set
+            {
+                if (_exchange.Timezone != value)
+                {
+                    _exchange.Timezone = value;
+                    RaisePropertyChanged(nameof(Timezone));
+                    ValidateProperty(_exchange, nameof(Timezone));
+                }
+            }
+        }
+
+        public bool IsAlwaysOpen
+        {
+            get { return _exchange.IsAlwaysOpen; }
+            set
+            {
+                if (_exchange.IsAlwaysOpen != value)
+                {
+                    _exchange.IsAlwaysOpen = value;
+                    RaisePropertyChanged(nameof(IsAlwaysOpen));
+                    ValidateProperty(_exchange, nameof(IsAlwaysOpen));
                 }
             }
         }