+++ OzzTradeDiary.NET/OzzTradeDiary/Helpers/QueryParameters.cs
@@ -37,12 +37,12 @@ namespace TD.Helpers
         /// Gets or sets the current page number.
         /// </summary>
         public int Page { get; set; }
-        
+
         /// <summary>
         /// Gets or sets the number of items per page.
         /// </summary>
         public int PageSize { get; set; }
-        
+
         /// <summary>
         /// Gets or sets the general purpose search string.
         /// </summary>